lots of cleanup, added get() to storage plugins, added ProxyStream for storage plugins

+++ plugins/FileStorage.py
@@ -7,6 +7,12 @@ from shutil import copy
 from yapsy.IPlugin import IPlugin
 import plugins
 
+try:
+    from cacus import common
+except ImportError:
+    import sys
+    sys.path.append('../..')
+    from cacus import common
 log = logging.getLogger('cacus.file_storage')
 
 
@@ -18,7 +24,16 @@ class FileStorage(plugins.IStoragePlugin):
             os.makedirs(self.root)
 
     def delete(self, key):
-        os.unlink(os.path.join(self.root, key))
+        try:
+            fname = os.path.join(self.root, key)
+            if not os.path.isfile(fname):
+                return common.Result('NOT_FOUND', 'File not found')
+            else:
+                os.unlink(os.path.join(self.root, key))
+        except Exception as e:
+            log.error("Cannot delete file %s: %s", key, e)
+            return common.Result('ERROR', e)
+        return common.Result('OK')
 
     def put(self, key, filename=None, file=None):
         #TODO: hashdir mb?
@@ -30,7 +45,7 @@ class FileStorage(plugins.IStoragePlugin):
                 os.makedirs(storage_dir)
             except Exception as e:
                 log.critical("Cannot create path for given key '%s': %s", key, e)
-                return None
+                return common.Result('ERROR', e)
 
         if filename:
             log.debug("Uploading from %s to %s", filename, storage_path)
@@ -38,7 +53,7 @@ class FileStorage(plugins.IStoragePlugin):
                 copy(filename, storage_path)
             except Exception as e:
                 log.critical("Cannot upload file: %s", e)
-                return None
+                return common.Result('ERROR', e)
         elif file:
             log.debug("Uploading from <stream> to %s", storage_path)
             try:
@@ -50,10 +65,25 @@ class FileStorage(plugins.IStoragePlugin):
                 file.seek(old_pos)
             except Exception as e:
                 log.critical("Cannot upload file: %s", e)
-                return None
+                return common.Result('ERROR', e)
 
-        return storage_key
+        return common.Result('OK', 'OK', storage_key)
 
 
-    def get(self, key):
-        return os.path.join(self.root, key)
+    def get(self, key, stream):
+        try:
+            fname = os.path.join(self.root, key)
+            if not os.path.isfile(fname):
+                return common.Result('NOT_FOUND', 'File not found')
+            else:
+                f = open(fname, 'r')
+        except Exception as e:
+            log.error("Cannot open file %s: %s", key, e)
+            return common.Result('ERROR', e)
+        for chunk in iter(lambda: f.read(4*1024*1024), b''):
+            try:
+                stream.write(chunk)
+            except IOError:
+                # remote side closed connection
+                break
+        return common.Result('OK')
